+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 * [NixOS/Features/Utilities] Add `fzf`.
+* [home-manager/Git] Configure for autosquash in rebase and add a `git fixup`
+    alias.
 * [Zsh/Git] Add aliases for `git worktree`.
 
 ### Removed